Add Dockerfile and continuous delivery (#5)
* Add Dockerfile naively

* Make Dockerfile self-contained

The shared Docker images do all sorts of Buf-specific stuff (pulling in
tokens, downloading commonly-used private packages, etc). They also make
this look much more complex than it is. Instead, build the demo service
with a pretty vanilla 2-stage build.

* Redirect / to landing page

Unless someone's hitting the RPC endpoints, redirect to the landing
page.

* Extend timeouts for streaming

Since we're exposing a bidi stream, give clients enough time to actually
use it.

* Deploy from Github actions

Build and deploy the Cloud Run app every time we land a commit on main
or manually trigger the workflow.

* Fix license checking

* Trim common leading and trailing punctuation

The regexp substitutions are much better if we trim common punctuation.

* Fix headline response

We're using this response on the landing page, but it's wrong! Because
we lowercase input before matching, we can't use `%s` as the first word
in the response.

* Fix ellipses

Rather than `... `, use `...` (no trailing space). I don't know why, but
this was driving me nuts.

* commit

Co-authored-by: Alan Chiu <awlchiiu@gmail.com>
Co-authored-by: Akshay Shah <akshay@akshayshah.org>
Co-authored-by: bufdev <bufdev-github@buf.build>
Co-authored-by: bufdev <4228796+bufdev@users.noreply.github.com>