Make protobuf package avoid reserved words
As part of moving this repository, we should change the protobuf
package structure. First, we drop the `buf` prefix. Second, we change
from `connect` to `connectrpc`: this matches the GitHub organization and
name, and it avoids problems in C++ (where "connect" isn't a valid
namespace). Finally, we drop "demo": it's needless categorization, and
it makes HTTP paths uncomfortably long.